requirements.txt
anyio==4.3.0
certifi==2024.2.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.6
pathlib==1.0.1
python-dotenv==1.0.1
python-telegram-bot==20.3
sniffio==1.3.1
